SNIG idx, css & js mod
May 25 10:31 PM

+++ !snig.js
@@ -56,15 +56,15 @@
 	
 	function PLAYwk(pl,pNO,wk) {
 		if ( document.getElementById("G"+pNO+"wk"+wk).style.background == "lime" ) {
-			alert(pl+" is NOT playing wk "+wk+"!");
+			//alert(pl+" is NOT playing wk "+wk+"!");
 			document.getElementById("G"+pNO+"wk"+wk).style.background = "slategray";	
 			localStorage.setItem("G"+pNO+"playWK"+wk,"");
 			
 			if ( wk==1 ) {
-				alert("PLAYwk1 array includes "+pl+". Will REMOVE from array!");
-				var index = PLAYwk1.indexOf(pl);		alert("index of "+pl+" = "+index+"\nPLAY = "+PLAYwk1);
+				//alert("PLAYwk1 array includes "+pl+". Will REMOVE from array!");
+				var index = PLAYwk1.indexOf(pl);		//alert("index of "+pl+" = "+index+"\nPLAY = "+PLAYwk1);
 				PLAYwk1.splice(index, 1);				/* <<< HELP! (July 15/23 */		// < **!
-				alert("Post splice PLAYwk1 = "+PLAYwk1);
+				//alert("Post splice PLAYwk1 = "+PLAYwk1);
 			}	
 		}
 		else { 
@@ -74,7 +74,7 @@
 			
 			if ( wk == 1 ) {
 				PLAYwk1.push(pl); 			//alert("PLAY array = "+PLAY);
-				alert("'PLAYwk1'array = "+PLAYwk1);
+				//alert("'PLAYwk1'array = "+PLAYwk1);
 			}
 		}		
 		
@@ -82,10 +82,10 @@
 	}
 	
 	function locStor() {
-		alert("localStorage:\n\n"+
+		/* alert("localStorage:\n\n"+
 			  "lS 'G3wk0' = "+localStorage.getItem('G3playWK0')+
 			  "\nlS 'G1wk3' = "+localStorage.getItem('G1playWK3')+
-			  "\n\nG1wk1: "+localStorage.getItem('G1playWK1')+" | G2wk1: "+localStorage.getItem('G2playWK1')+" | G3wk1: "+localStorage.getItem('G3playWK1')+" | G4wk1: "+localStorage.getItem('G4playWK1')+" | G5wk1: "+localStorage.getItem('G5playWK1') );
+			  "\n\nG1wk1: "+localStorage.getItem('G1playWK1')+" | G2wk1: "+localStorage.getItem('G2playWK1')+" | G3wk1: "+localStorage.getItem('G3playWK1')+" | G4wk1: "+localStorage.getItem('G4playWK1')+" | G5wk1: "+localStorage.getItem('G5playWK1') ); */
 			  		
 /*		for (let i = 1; i < 6; i++) {	  
 			if ( localStorage.getItem('G'+i+'playWK1') !== "" ) {
@@ -94,18 +94,32 @@
 		}
 */	
 		for (let w = 0; w < 5; w++) {	  
-			for (let i = 1; i < 11; i++) {	  
+			for (let i = 1; i < 17; i++) {	  
 				if ( (localStorage.getItem('G'+i+'playWK'+w) != "") &&  (localStorage.getItem('G'+i+'playWK'+w) != null) ) {
 					document.getElementById("G"+i+"wk"+w).style.background = "green";
 				}
 			}
 		}
-			  
+
+		document.getElementById("mySidenav").style.width = "0px";
+		
 	}
 	
 	function clearLS () {		
-			alert("G1playWK0 PRE clear = "+localStorage.getItem('G1playWK0'));
-			//localStorage.setItem("lastPLAY","");
-			localStorage.clear();
-			alert("G1playWK0 post clear = "+localStorage.getItem('G1playWK0'));
+			//alert("G1playWK0 PRE clear = "+localStorage.getItem('G1playWK0'));
+			
+		  let reply;
+		  if (confirm("Are you SURE you want to CLEAR 'localStorage'?") == true) {
+		    reply = "YES!";
+		  } else {
+		    reply = "NO!";
+		  }
+  		  alert(reply);
+			  
+		  if ( reply=="YES!") {			  
+				//localStorage.setItem("lastPLAY","");
+				localStorage.clear();
+				//alert("G1playWK0 post clear = "+localStorage.getItem('G1playWK0'));
+				alert("'localStorage' has been CLEARED!");
+		  }
 	}
